Added throwing tiles
Tiles can be thrown via left click when in tile throwing mode. There are two buttons on the bottom of the screen that shows what mode the player is in and allows the to switch between the two. Also updated the end turn button to show when its the ai's turn by disabling