Added visible player stats via left shift
Pressing left shift brings up the players stats, pressing it again removes it. The stats are updated in real time and turn green if they are above base and red if they are below base. TODO: have a base health and a current health ie 78/100 health; make the element mods read 100% instead of 1