Refactored items and V1 of show item effects
Passive Items are actual items now! Instead of just modifying a stat, Items are objects with a name, a sprite, and a description. In item manager based on what random item is generated, the effect is applied. Item effect is displayed in a popup to the right of the chest. Popup background is static low res ui sprite, will update to fancy re-sizing one later aswell as making text look better.